First pass on project structure, should enable feature build out
