+++ lib/core/es6.js
