+++ lib/core/tree.js
