+++ lib/six.js
@@ -6,7 +6,7 @@ if (typeof exports === 'object' && typeof define !== 'function') {
 define(function (require, exports) {
   var fs = require("fs");
   var path = require("path");
-  var rewrite = require("./rewriter").rewrite;
+  var rewrite = require("./core/rewriter").rewrite;
   var run = exports.run = function (code, options) {
       var mainModule = require.main;
       if (!options)
